refactor: isolate turn route resolution owner
